+++ NESseract/NESseract.Core/Cpu/Operations/ANDOperation.cs
@@ -12,7 +12,7 @@ namespace NESseract.Core.Cpu.Operations
          var result = operationValue & registers.A;
 
          registers.N_NegativeFlag = (byte)((result & 0x80) >> 7);
-         registers.Z_ZeroFlag = (byte)result == 0 ? 1 : 0;
+         registers.Z_ZeroFlag = (byte)((byte)result == 0 ? 1 : 0);
 
          registers.A = (byte)result;
