+++ NESseract/NESseract.Core/Cpu/Operations/CMPOperation.cs
@@ -9,12 +9,12 @@ namespace NESseract.Core.Cpu.Operations
       {
          var operationValue = addressingMode.GetValue(memory, registers, operand1, operand2, out bool pageBoundaryCrossed);
 
-         registers.C_CarryFlag = registers.A >= operationValue ? 1 : 0;
+         registers.C_CarryFlag = (byte)(registers.A >= operationValue ? 1 : 0);
 
          var result = registers.A - operationValue;
 
          registers.N_NegativeFlag = (byte)((result & 0x80) >> 7);
-         registers.Z_ZeroFlag = (byte)result == 0 ? 1 : 0;
+         registers.Z_ZeroFlag = (byte)((byte)result == 0 ? 1 : 0);
 
          return (byte)(opCodeDefinition.ExecutionCycles + (opCodeDefinition.AddExecutionCycleOnPageBoundaryCross && pageBoundaryCrossed ? 1 : 0));
       }
