+++ NESseract/NESseract.Core/Cpu/Operations/INCOperation.cs
@@ -13,7 +13,7 @@ namespace NESseract.Core.Cpu.Operations
          var result = operationValue + 1;
 
          registers.N_NegativeFlag = (byte)((result & 0x80) >> 7);
-         registers.Z_ZeroFlag = (byte)result == 0 ? 1 : 0;
+         registers.Z_ZeroFlag = (byte)((byte)result == 0 ? 1 : 0);
 
          addressingMode.SetValue(memory, registers, operationAddress, (byte)result);
