+++ NESseract/NESseract.Core/Cpu/Operations/ROROperation.cs
@@ -14,7 +14,7 @@ namespace NESseract.Core.Cpu.Operations
 
          registers.N_NegativeFlag = (byte)((result & 0x80) >> 7);
          registers.C_CarryFlag = (byte)(operationValue & 0x01);
-         registers.Z_ZeroFlag = (byte)result == 0x00 ? 1 : 0;
+         registers.Z_ZeroFlag = (byte)((byte)result == 0x00 ? 1 : 0);
 
          addressingMode.SetValue(memory, registers, operationAddress, (byte)result);
