+++ NESseract/NESseract.Core/Cpu/Operations/TAYOperation.cs
@@ -8,7 +8,7 @@ namespace NESseract.Core.Cpu.Operations
       public byte Execute(OpCodeDefinition opCodeDefinition, IAddressingMode addressingMode, CPUMemory memory, CPURegisters registers, byte operand1, byte operand2)
       {
          registers.N_NegativeFlag = (byte)((registers.A & 0x80) >> 7);
-         registers.Z_ZeroFlag = registers.A == 0 ? 1 : 0;
+         registers.Z_ZeroFlag = (byte)(registers.A == 0 ? 1 : 0);
 
          registers.Y = registers.A;
